Include the initial graph comparison notebook
Include a notebook to load the same dataset as used within the GMN
papers and run it through the backtracking algorithm to find the
maximum common subgraph (MCS) between the learning data structure
and the canoncial form. Based upon the MCS, then calculate the
jaccard index and display within a heatmap.